+++ src/Timestampable/Mapping/Driver/Attribute.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Gedmo\Timestampable\Mapping\Driver;
+
+use Gedmo\Mapping\Driver\AttributeDriverInterface;
+
+/**
+ * This is an attribute mapping driver for Timestampable
+ * behavioral extension. Used for extraction of extended
+ * metadata from attributes specifically for Timestampable
+ * extension.
+ *
+ * @author Kevin Mian Kraiker <kevin.mian@gmail.com>
+ * @license MIT License (http://www.opensource.org/licenses/mit-license.php)
+ *
+ * @internal
+ */
+final class Attribute extends Annotation implements AttributeDriverInterface
+{
+}