+++ tests/Gedmo/Timestampable/Fixture/MappedSupperClass.php
@@ -2,12 +2,14 @@
 
 namespace Gedmo\Tests\Timestampable\Fixture;
 
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
 
 /**
  * @ORM\MappedSuperclass
  */
+#[ORM\MappedSuperclass]
 class MappedSupperClass
 {
     /**
@@ -17,6 +19,9 @@ class MappedSupperClass
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="AUTO")
      */
+    #[ORM\Column(name: 'id', type: Types::INTEGER)]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'AUTO')]
     protected $id;
 
     /**
@@ -24,6 +29,7 @@ class MappedSupperClass
      *
      * @Gedmo\Locale
      */
+    #[Gedmo\Locale]
     protected $locale;
 
     /**
@@ -32,6 +38,8 @@ class MappedSupperClass
      * @Gedmo\Translatable
      * @ORM\Column(name="name", type="string", length=191)
      */
+    #[Gedmo\Translatable]
+    #[ORM\Column(name: 'name', type: Types::STRING, length: 191)]
     protected $name;
 
     /**
@@ -40,6 +48,8 @@ class MappedSupperClass
      * @ORM\Column(name="created_at", type="datetime")
      * @Gedmo\Timestampable(on="create")
      */
+    #[ORM\Column(name: 'created_at', type: Types::DATETIME_MUTABLE)]
+    #[Gedmo\Timestampable(on: 'create')]
     protected $createdAt;
 
     /**