+++ tests/Gedmo/Timestampable/Fixture/SupperClassExtension.php
@@ -8,12 +8,15 @@ use Gedmo\Mapping\Annotation as Gedmo;
 /**
  * @ORM\Entity
  */
+#[ORM\Entity]
 class SupperClassExtension extends MappedSupperClass
 {
     /**
      * @ORM\Column(length=128)
      * @Gedmo\Translatable
      */
+    #[ORM\Column(length: 128)]
+    #[Gedmo\Translatable]
     private $title;
 
     public function setTitle($title)