+++ tests/Gedmo/Timestampable/Fixture/TitledArticle.php
@@ -2,6 +2,7 @@
 
 namespace Gedmo\Tests\Timestampable\Fixture;
 
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
 use Gedmo\Timestampable\Timestampable;
@@ -9,6 +10,7 @@ use Gedmo\Timestampable\Timestampable;
 /**
  * @ORM\Entity
  */
+#[ORM\Entity]
 class TitledArticle implements Timestampable
 {
     /**
@@ -16,21 +18,27 @@ class TitledArticle implements Timestampable
      * @ORM\GeneratedValue
      * @ORM\Column(type="integer")
      */
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: Types::INTEGER)]
     private $id;
 
     /**
      * @ORM\Column(name="title", type="string", length=128)
      */
+    #[ORM\Column(name: 'title', type: Types::STRING, length: 128)]
     private $title;
 
     /**
      * @ORM\Column(name="text", type="string", length=128)
      */
+    #[ORM\Column(name: 'text', type: Types::STRING, length: 128)]
     private $text;
 
     /**
      * @ORM\Column(name="state", type="string", length=128)
      */
+    #[ORM\Column(name: 'state', type: Types::STRING, length: 128)]
     private $state;
 
     /**
@@ -39,7 +47,9 @@ class TitledArticle implements Timestampable
      * @ORM\Column(name="chtext", type="datetime", nullable=true)
      * @Gedmo\Timestampable(on="change", field="text")
      */
-    private $chtext;
+    #[ORM\Column(name: 'chtext', type: Types::DATETIME_MUTABLE, nullable: true)]
+    #[Gedmo\Timestampable(on: 'change', field: 'text')]
+    private $chText;
 
     /**
      * @var \DateTime
@@ -47,7 +57,9 @@ class TitledArticle implements Timestampable
      * @ORM\Column(name="chtitle", type="datetime", nullable=true)
      * @Gedmo\Timestampable(on="change", field="title")
      */
-    private $chtitle;
+    #[ORM\Column(name: 'chtitle', type: Types::DATETIME_MUTABLE, nullable: true)]
+    #[Gedmo\Timestampable(on: 'change', field: 'title')]
+    private $chTitle;
 
     /**
      * @var \DateTime
@@ -55,38 +67,40 @@ class TitledArticle implements Timestampable
      * @ORM\Column(name="closed", type="datetime", nullable=true)
      * @Gedmo\Timestampable(on="change", field="state", value={"Published", "Closed"})
      */
+    #[ORM\Column(name: 'closed', type: TYPES::DATETIME_MUTABLE, nullable: true)]
+    #[Gedmo\Timestampable(on: 'change', field: 'state', value: ['Published', 'Closed'])]
     private $closed;
 
     /**
-     * @param \DateTime $chtext
+     * @param \DateTime $chText
      */
-    public function setChtext($chtext)
+    public function setChText($chText)
     {
-        $this->chtext = $chtext;
+        $this->chText = $chText;
     }
 
     /**
      * @return \DateTime
      */
-    public function getChtext()
+    public function getChText()
     {
-        return $this->chtext;
+        return $this->chText;
     }
 
     /**
-     * @param \DateTime $chtitle
+     * @param \DateTime $chTitle
      */
-    public function setChtitle($chtitle)
+    public function setChTitle($chTitle)
     {
-        $this->chtitle = $chtitle;
+        $this->chTitle = $chTitle;
     }
 
     /**
      * @return \DateTime
      */
-    public function getChtitle()
+    public function getChTitle()
     {
-        return $this->chtitle;
+        return $this->chTitle;
     }
 
     /**