+++ tests/Gedmo/Timestampable/Fixture/UsingTrait.php
@@ -2,12 +2,14 @@
 
 namespace Gedmo\Tests\Timestampable\Fixture;
 
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Timestampable\Traits\TimestampableEntity;
 
 /**
  * @ORM\Entity
  */
+#[ORM\Entity]
 class UsingTrait
 {
     /*
@@ -21,11 +23,15 @@ class UsingTrait
      * @ORM\GeneratedValue
      * @ORM\Column(type="integer")
      */
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: Types::INTEGER)]
     private $id;
 
     /**
      * @ORM\Column(length=128)
      */
+    #[ORM\Column(length: 128)]
     private $title;
 
     public function getId()