+++ tests/Gedmo/Timestampable/Fixture/WithoutInterface.php
@@ -2,32 +2,46 @@
 
 namespace Gedmo\Tests\Timestampable\Fixture;
 
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
 
 /**
  * @ORM\Entity
  */
+#[ORM\Entity]
 class WithoutInterface
 {
-    /** @ORM\Id @ORM\GeneratedValue @ORM\Column(type="integer") */
+    /**
+     * @ORM\Id
+     * @ORM\GeneratedValue
+     * @ORM\Column(type="integer")
+     */
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: Types::INTEGER)]
     private $id;
 
     /**
      * @ORM\Column(type="string", length=128)
      */
+    #[ORM\Column(type: Types::STRING, length: 128)]
     private $title;
 
     /**
      * @Gedmo\Timestampable(on="create")
      * @ORM\Column(type="date")
      */
+    #[Gedmo\Timestampable(on: 'create')]
+    #[ORM\Column(type: Types::DATE_MUTABLE)]
     private $created;
 
     /**
      * @ORM\Column(type="datetime")
      * @Gedmo\Timestampable(on="update")
      */
+    #[ORM\Column(type: Types::DATETIME_MUTABLE)]
+    #[Gedmo\Timestampable(on: 'update')]
     private $updated;
 
     public function getId()