+++ CHANGELOG.md
@@ -19,6 +19,8 @@ a release.
 ---
 
 ## [Unreleased]
+### Added
+- Timestampable: Support to use annotations as attributes on PHP >= 8.0.
 
 ## [3.3.1] - 2021-11-18
 ### Fixed